metayer environment variables and options

metayer environment variables and options

User facing configuration

config.yml

Package configuration is specified in config.yml, which is designed to mimic the behavior of the config package. As such, there are currently three top-level sections: default, testing, and jupyter. Our extension of config::get enables YAML references, and the pattern is that top-level sections are meant to inherit from default:

Note that there are two YAML tags in use, !optenv and !with_env.

!optenv

optenv is used to evaluate simple R expressions, like logger::INFO; the idea is to simplify the specification of objects or functions.

!with_env

with_env allows strings to contain environment variables.

onLoad tasks

When metayer is loaded, it populates several options and sets up the loggers. Portions of config.yml are named accordingly

logger

format

A string specification for the logging format passed to logger::log_format

logfile

The location of a log file. This will become the log_appender at namespace global and index one.

appenders

A list of appenders that will populate subsequent indexes of the global namespace.

threshold

The log threshold, e.g., logger::INFO

options

cli.default_handler

The same cli.default_handler as in the cli package. We set this to cli_nullity_handler in order to enjoy null visibility in cli output.

mty.cli_null

The substitution made for NULL values in cli output.

knitr.chunk.metayer_hook

A knitr option. Setting this enables sure_plot to produce more uniform plotting results through a custom knitr hook (specified in .onLoad).

knitr.chunk.*

knitr chunk options

knitr.chunk.dev

The default device used when knitting.

knitr.chunk.comment

A cosmetic change for R output.

knitr.chunk.class.output

A cosmetic change for imbuing R output with class tags suitable for CSS.

knitr.package.*

knitr package options

knitr.package.root.dir

The directory in which knitr will expect to find references. pkgdown and rmarkdown behavior can be quite sensitive to this.

mty.fig.*

with_grdev default parameters. See the function reference.

display

ippx, ippy

Inches per pixel. Currently unused although likley to be useful when creating interactive devices–e.g. X11–and specifying dimensions in pixels rather than inches.

rdoc_config

This section serves as the default YAML header for new jupyter notebooks. With this, one really only needs to specify a title when creating new documents.

all_doc

Shared across all rmarkdown document types, this is backreferenced as pandoc_shared

output

The usual jupyter YAML content.

rmall_exclusions

A list of functions to blacklist when running rm.all.