Cli wrapper logic

wrap_factory uses the body of this function as a building block. It does this using 'substitute', replacing the cmd and level parameters with their known expressions.

Usage

wrap_cli_body(cmd_call, level, .caller_env = caller_env())

Arguments

cmd_call

a call, the cli operation that we're wrapping

level

the logger level to associated with the cmd

.caller_env

the caller environment in which to evaluate cmd_call

Details

Note, the awkward substitute construction:

local_env <- new_environment(list(cmd = cmd)) injection <- lapply( body(wrap_cli_body), function(s) do.call(substitute, list(s, local_env)) ) %>% as.call()